Fixed a few files to compile again after his latest changes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15691 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/aws/awsmled.cpp
@@ -371,7 +371,7 @@ bool awsMultiLineEdit::HandleEvent (iEvent &Event)
     {
       csKeyEventData eventData (&Event);
       utf32_char Char[2];
-      size_t composedSize;
+      int composedSize;
 
       if (composer->HandleKey (eventData, Char, 
 	sizeof (Char) / sizeof (utf32_char), &composedSize) != csComposeNoChar)
